Change decrypt to return Buffer instead of string.
Currently encrypt takes in a Buffer while decrypt returns a utf8-string decoded from Buffer.
This means if a buffer containing non-utf8 string is encrypted, decrypt won't return the
correct value. This commit changes the return type to Buffer and relies on the caller whom
knows the type to interpret the bytes.